requirements.txt
streamlit
wikipedia
Pillow
requests
